django-web/webapp/medbot2_updated/medbot2/requirements.txt
langchain
langchain-groq
langchain-community
langchain-experimental
pandas
python-dotenv
pymongo
streamlit
sentence_transformers

